Regression(r238330) ProcessSwap.WebInspector API test is a flaky crash in debug https://bugs.webkit.org/show_bug.cgi?id=191812
Patch by Joseph Pecoraro <pecoraro@apple.com> on 2018-11-18
Reviewed by Chris Dumez.

* WebProcess/WebPage/WebInspector.cpp:
(WebKit::WebInspector::setFrontendConnection):
It can be possible to receive a frontend connection multiple times
if another WebProcess opened the inspector and this WebProcess got
swapped in before the InspectorProcess notified the WebProcess of
a connection to use. It is always best to use the latest connection
we have received, since any others would have been invalidated.

Identifier: 206545@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@238369 268f45cc-cd09-0410-ab3c-d52691b4dbfc