[Flatpak] Do not disable signals when running "usual" commands https://bugs.webkit.org/show_bug.cgi?id=190679
Patch by Thibault Saunier <tsaunier@igalia.com> on 2018-11-22
Reviewed by Philippe Normand.

Otherwise <ctr>c doesn't work while compiling. It is not the best fix
as there are cases we want to let only children receive signals, but
it is not clear how to determine when (when running gdb, we still
disable those).

* flatpak/flatpakutils.py:
(WebkitFlatpak.run_in_sandbox):

Identifier: 206613@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@238444 268f45cc-cd09-0410-ab3c-d52691b4dbfc