Better parsing of comments in generate-message*.py https://bugs.webkit.org/show_bug.cgi?id=191866 <rdar://problem/46189563>
Reviewed by Chris Dumez.

The script parsing the *.messages.in files would treat a line starting
with '#' as a comment, but not a line starting with '<whitespace>#'.
This means that jamming a '#' right in front of the first character of
a message definition (as opposed to the beginning of a line) will have
no effect and the line will get treated just the same as a
non-commented line. Fix this by trimming all white space from the
beginning and ending of the line before processing it.

* Scripts/webkit/parser.py:
(parse):
* Scripts/webkit/test-messages.in:


Identifier: 206735@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@238577 268f45cc-cd09-0410-ab3c-d52691b4dbfc