[Cocoa] Streamline sandbox code a tiny bit https://bugs.webkit.org/show_bug.cgi?id=192299
Reviewed by Anders Carlsson.

* Shared/mac/SandboxUtilities.mm:
(WebKit::processHasContainer): Removed separate function, just use the
pathForProcessContainer instead of repeating logic. This costs a tiny
bit of memory allocation overhead the first time this function is called,
but in return we can eliminate a second copy of the code to call
sandbox_container_path_for_pid.


Identifier: 206941@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@238797 268f45cc-cd09-0410-ab3c-d52691b4dbfc