[Cocoa] Check descriptor types in createMessageDecoder https://bugs.webkit.org/show_bug.cgi?id=192302
Reviewed by Chris Dumez.

Not sure how to construct a test case for Connection, so not adding any
tests at this time.

* Platform/IPC/mac/ConnectionMac.mm:
(IPC::createMessageDecoder): Check descriptor types and return nullptr if they
are incorrect. Retained the assertions to help with debugging.
(IPC::Connection::receiveSourceEventHandler): Ignore the message if no decoder
is created, which means the message is invalid.

Identifier: 206943@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@238799 268f45cc-cd09-0410-ab3c-d52691b4dbfc